vfs: Add support to skip trash if 's' attribute is set
This change adds support for the FlagSkipTrash in the VFS and meta portions
of the code and maps to FS_SECRM_FL / secure removal / 's' in chattr. Setting
the attribute on a file will prevent the file from being sent to trash. New files
and sub-directories will inherit the flag if the directory has the flag set.

During a rename:

* The flag will be inherited from the source file but not the source directory
* The flag will be inherited from the destination directory or the destination
  file if the source file replaces an existing file